+++ src/main/java/ru/hollowhorizon/hollowengine/client/gui/DashboardScreen.kt
@@ -54,7 +54,7 @@ object DashBoardScreen : ImGuiScreen() {
 
     fun imageButton(image: String, size: Float): Boolean {
         ImGui.pushID(image)
-        val isClicked = ImGui.imageButton("hollowengine:textures/gui/icons/$image.png".rl.toTexture().id, size, size)
+        val isClicked = ImGui.imageButton("hollowengine:textures/gui/icons/$image.png".rl.toTexture().id.toLong(), size, size)
         ImGui.pushStyleVar(ImGuiStyleVar.PopupBorderSize, 3f)
         if (ImGui.isItemHovered()) ImGui.setTooltip(
             //? if >1.20.1 {
